requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1
streamlit
requests

